requirements.txt
pandas==2.2.2
numpy==1.26.4
selenium==4.23.1
beautifulsoup4==4.12.3
sqlalchemy==2.0.31
scikit-learn==1.5.1
joblib==1.4.2
streamlit==1.37.0
plotly==5.22.0
python-dotenv==1.0.1
webdriver-manager==4.0.2
psycopg==3.1.19
